requirements.txt
streamlit
langchain
pdfminer.six
sentence-transformers
faiss-cpu
langchain-community